backend-kil/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.4,<5.0
django-environ>=0.11,<1.0
celery[redis]>=5.4,<6.0
redis>=5.0,<6.0
django-redis>=5.4,<6.0
Pillow>=10.3,<11.0
drf-spectacular>=0.27,<1.0
django-filter>=24.2,<25.0
whitenoise>=6.7,<7.0
django-celery-beat>=2.6,<3.0
requests>=2.32,<3.0
ipython>=8.24,<9.0
black>=24.4,<25.0
ruff>=0.4,<1.0
gunicorn>=22.0,<23.0
psycopg2-binary>=2.9,<3.0
